clicktest: add support for floating point range expressions
Previously, I added integer range expressions. In this commit, I expand
them to "numeric range expressions", which support integer and float
ranges. The change in format is pretty minimal:  if any of the numbers
in the format contains a decimal point, floats and integers are
accepted; otherwise only integers are accepted (as before). Here is an
example of the float variant of each type:
  {{~ 5.0 - 15 }}   # note: equivalent to {{~ 5 - 15.0 }}
  {{~ 100 +- 5.0 }}
  {{~ 100 +- 5.0% }}

Reviewed-by: Yoann Desmouceaux <ydesmouc@meraki.com>